Refactor and rename Expression Validation, Parsing, and Evaluation classes
