+++ core/src/test/scala/org/broadinstitute/dsde/rawls/jobexec/MethodConfigResolverSpec.scala
@@ -129,7 +129,7 @@ class MethodConfigResolverSpec extends WordSpecLike with Matchers with TestDrive
         case scala.util.Failure(exception) =>
           DBIO.failed(exception)
         case scala.util.Success(methodInputs) =>
-          MethodConfigResolver.resolveInputsForEntities(workspaceContext, methodInputs, entityRecs, dataAccess)
+          MethodConfigResolver.evaluateInputExpressions(workspaceContext, methodInputs, entityRecs, dataAccess)
       }
     }
   }
